sw/webapp/requirements.txt
Django==1.7
django-registration-redux==1.2
pyserial==3.2.1
wheel==0.24.0
wsgiref==0.1.2
